[Refactor] 클래스 이름 및 value 타입 변경
- StatisticDto를 StatisticResponse로 클래스 이름 변경 (도메인 + 기능 + Request/Response 네이밍 규칙에 따라 통일)
- 기존 StatisticResponse를 StatisticQueryResponse로 클래스 이름 변경
- value 타입을 double에서 Long으로 변경 (0.0 표기 문제 해결)
