+++ src/main/java/com/allclear/socialhub/post/common/view/repository/PostViewRepository.java
@@ -1,6 +1,6 @@
 package com.allclear.socialhub.post.common.view.repository;
 
-import com.allclear.socialhub.post.common.response.StatisticResponse;
+import com.allclear.socialhub.post.common.response.StatisticQueryResponse;
 import com.allclear.socialhub.post.common.view.domain.PostView;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -17,7 +17,7 @@ public interface PostViewRepository extends JpaRepository<PostView, Long> {
             "AND DATE(pv.createdAt) BETWEEN :start AND :end " +
             "GROUP BY DATE_FORMAT(pv.createdAt, '%Y-%m-%d') " +
             "ORDER BY time ASC")
-    List<StatisticResponse> findDailyStatisticByPostIds(
+    List<StatisticQueryResponse> findDailyStatisticByPostIds(
             @Param("postIds") List<Long> postIds,
             @Param("start") LocalDate start,
             @Param("end") LocalDate end);
