+++ src/main/java/com/allclear/socialhub/post/service/StatisticService.java
@@ -2,7 +2,7 @@ package com.allclear.socialhub.post.service;
 
 import com.allclear.socialhub.post.domain.StatisticType;
 import com.allclear.socialhub.post.domain.StatisticValue;
-import com.allclear.socialhub.post.dto.StatisticDto;
+import com.allclear.socialhub.post.dto.StatisticResponse;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
@@ -11,8 +11,8 @@ import java.util.List;
 @Service
 public interface StatisticService {
 
-    List<StatisticDto> getStatisticsDaily(String hashtag, StatisticType type, LocalDate start, LocalDate end, StatisticValue value);
+    List<StatisticResponse> getStatisticsDaily(String hashtag, StatisticType type, LocalDate start, LocalDate end, StatisticValue value);
 
-    void getStatisticsHourly(String hashtag, StatisticType type, LocalDate start, LocalDate end, StatisticValue value);
+    List<StatisticResponse> getStatisticsHourly(String hashtag, StatisticType type, LocalDate start, LocalDate end, StatisticValue value);
 
 }
